Render Counter
in addTodo,  got the value of the input by event.target.elements."name".value.   if that exists, push it to an array property of our app object.   And put the rendering of the template into a function, and call that function each time a new todo list is added.  because React works with a virtual DOM, the real DOM is not rerendered.