What Should I do button
rendered a button "what should i do" that randomly chooses one of the todos and alerts the user of the todo task.  acomplished by creating a function makeDecision that is called with the onClick event attribute on the button. it creates a random number  in the range of the items in the todo list and assigns a variable to one of the todos. then it alerts the user of the randomly chosen option.